Implement adding child controls and rework control traits some
The list of child controls is implemented using a ChildrenVec struct in
a RefCell. I tried using Cells or RefCell in ChildrenVec but that design
didn't work nearly as well.

Now the Control trait is implemented for any type that implements
SubControlRef to make it easier to implement (and still no macros
needed).

Also, convert some properties of a control to bit fields to save space.
Add a tab_index field and a function that sets tab indexes
consecutively.