Change event handling to use Any
I couldn't write

enum ButtonEvent {
    SubControl(&'a mut SubControlEvent),
    ...
}

because I'd have to add a lifetime to the ButtonEvent, and then I'd
have to add it to the ButtonData struct even though no references are
stored. And adding a lifetime to ButtonData is unacceptable. Using Any,
there will only be one list of event handlers, so it is noticeably
easier to implement a control because it reuses the SubControl event
handler implementation, and it's roughly the same handling events.

Downcasting Any should be quite fast, as it is only a virtual function
call to get the u64 TypeId, then comparing it to a constant (after which
it reinterprets the pointer, which should be free).